Apply automatic clippy fixes in runtime and FFI (#5426)
* Apply automatic clippy fixes in runtime and FFI

* fix: restore test-referenced items clippy --fix removed (perry-runtime)

clippy --fix dropped `ObjectHeader` and `js_object_get_field_by_name_f64`
from the `use crate::object::{...}` in node_stream.rs because they are only
referenced by `#[cfg(test)]` code (node_stream_state_tests.rs et al. via
`use super::*`). That broke `cargo test -p perry-runtime` with E0425. Restore
them under a `#[cfg(test)]` import so the non-test build stays clippy-clean.


---------

+++ crates/perry-runtime/src/array/splice_slice.rs
@@ -113,7 +113,7 @@ pub extern "C" fn js_array_splice(
         }
 
         // Calculate new length
-        let new_len = (len as u32 - actual_delete + items_count) as u32;
+        let new_len = len as u32 - actual_delete + items_count;
 
         // Grow array if needed
         let arr = if new_len > (*arr).capacity {
