+++ crates/perry-runtime/src/collection_iter.rs
@@ -178,7 +178,7 @@ pub(crate) fn require_callable(value: f64, name: &str) -> f64 {
 pub(crate) fn normalize_callable_value(value: f64) -> f64 {
     let raw = js_nanbox_get_pointer(value) & !0x7;
     if raw >= 0x10000 {
-        return crate::value::js_nanbox_pointer(raw as i64);
+        return crate::value::js_nanbox_pointer(raw);
     }
     value
 }
